Move the core reducers to the right folder
