+++ src/reducers/CoreWalletsReducer.js
@@ -3,7 +3,7 @@
 import type { EdgeCurrencyWallet } from 'edge-core-js'
 import { type Reducer, combineReducers } from 'redux'
 
-import { type Action } from '../../../types/reduxTypes.js'
+import { type Action } from '../types/reduxTypes.js'
 
 export type WalletsState = {
   byId: {
