+++ src/reducers/EdgeLoginReducer.js
@@ -3,7 +3,7 @@
 import type { EdgeLobby } from 'edge-core-js'
 import { type Reducer } from 'redux'
 
-import type { Action } from '../../types/reduxTypes.js'
+import type { Action } from '../types/reduxTypes.js'
 
 export type EdgeLoginState = {
   lobby: EdgeLobby | null,
