+++ src/reducers/RootReducer.js
@@ -2,11 +2,11 @@
 
 import { type Reducer, combineReducers } from 'redux'
 
-import { type CoreState, core } from '../modules/Core/reducer.js'
 import { type ExchangeRatesState, exchangeRates } from '../modules/ExchangeRates/reducer.js'
 import { type UiState, ui } from '../modules/UI/reducer.js'
 import { type Action } from '../types/reduxTypes.js'
 import { type ContactsState, contacts } from './ContactsReducer.js'
+import { type CoreState, core } from './CoreReducer.js'
 import { type CryptoExchangeState, cryptoExchange } from './CryptoExchangeReducer.js'
 import { type PermissionsState, permissions } from './PermissionsReducer.js'
